bugfix. Fix songtitle tiebreaker algorithm
Occasionally there will be overlap on song metadata
(bpm, note count, difficulty) and the OCR text
will not match neatly, so we have a tiebreaker
based on edit distance against the known song names
and artists.

The old algorithm died immediately if there was a tie,
when it should evaluate everything and pick the smallest
edit distances, and then fail if there's more than one
shortest edit distance.

Also added a test for this specific bug.